requirements.txt
aiohttp
async-timeout==3.0.0
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
idna==2.7
idna-ssl==1.1.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
multidict==4.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
yarl==1.2.6
